pr fixes 2
